fix(elasticsearch): Requested changes for ESSignalProcessor
